+++ backend/docker/git-askpass.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+case "$1" in
+  *Username*) printf '%s\n' 'x-access-token' ;;
+  *) printf '%s\n' "$ZIP_GITHUB_GIT_TOKEN" ;;
+esac